Migrate to Tailwind
Finally migrated from SCSS to Tailwind V4 after a couple of my friends
kept telling me how awesome it is and seeing it's popularity on the
internet. It's recognized as one of the best webdev stuff ever if that
makes any sense. It's not bad at all to be honest, it's actually pretty
nice once you get the hang of it, I'm liking it a lot so far no wonder
it's got such a high reputation. Still got a lot of code to clean up.
Also added some new plugins like Typography for Tailwind for beautiful
default typography styles whatever and `postcss-nested` so I can nest
styles just like in SCSS which I find really convenient.