Removed Link Component
I feel that the Link component is too overcomplicated and convoluted
just to style the anchor tag that links to the page that the user is
currently on. I'm sure there's a better and more practical way to do
this that I'm unaware of. The Link component was just too, odd I
suppose. I won't style the active link for now.