Readded Link Component
Readded the Link component which is far more straightforward and
readable this time round. I also created a separate component just for
the head because there's a lot of stuff in the head, this made the
`DefaultLayout.astro` file far more readable. I'm also trying out some
new fonts just for the sake of experimenting.